+++ lib/about.dart
@@ -1,4 +1,4 @@
-import 'package:doclense/MainDrawer.dart';
+import 'package:doclense/main_drawer.dart';
 import 'package:doclense/Providers/theme_provider.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
