+++ lib/image_view.dart
@@ -1,13 +1,14 @@
 import 'dart:io';
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:image_cropper/image_cropper.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:provider/provider.dart';
-import 'Filterview.dart';
-import 'Home.dart';
+
 import 'Providers/image_list.dart';
 import 'Providers/theme_provider.dart';
+import 'filter_view.dart';
+import 'home.dart';
 
 class Imageview extends StatefulWidget {
   final File file;
