+++ lib/settings.dart
@@ -1,12 +1,12 @@
 // import 'package:doclense/Constants/theme_constants.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:package_info/package_info.dart';
 import 'package:provider/provider.dart';
 import 'package:wiredash/wiredash.dart';
-import 'package:package_info/package_info.dart';
 
-import 'MainDrawer.dart';
 import 'Providers/theme_provider.dart';
+import 'main_drawer.dart';
 import 'setting_text.dart';
 
 class SettingsScreen extends StatefulWidget {
